minor changes to bytcode parsing

+++ src/main/java/me/lukas81298/decompiler/bytecode/atrr/impl/CodeAttribute.java
@@ -34,7 +34,7 @@ public class CodeAttribute extends AttributeData {
     @EqualsAndHashCode(callSuper = false)
     public final static class CodeItem {
         private final String id;
-        private final byte[] additionalData;
+        private final int[] additionalData;
     }
 
     @RequiredArgsConstructor
